auth-service/requirements.txt
flask
flask_sqlalchemy
psycopg2-binary
flask_cors
bcrypt
PyJWT
flasgger
pymongo
Flask-PyMongo
